requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
antiorm==1.2.1
anyio==3.7.1
APScheduler==3.6.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==4.2.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
db==0.1.1
db-sqlite3==0.0.1
exceptiongroup==1.1.2
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
multidict==6.0.4
outcome==1.2.0
pycparser==2.21
pyOpenSSL==23.2.0
PySocks==1.7.1
python-telegram-bot==13.7
pytz==2023.3
requests==2.31.0
selenium==4.0.0b4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
telepot==12.7
tornado==6.3.2
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
urllib3-secure-extra==0.1.0
wsproto==1.2.0
yarl==1.9.2
